scorecard/pyproject.toml
[project]
name = "scorecard"
version = "0.1.0"
description = "update the scorecard figures monthly"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "az-pmp-utils",
    "google-api-python-client>=2.191.0",
    "google-auth>=2.48.0",
    "polars>=1.38.1",
    "sendgrid>=6.12.5",
]

[tool.basedpyright]
typeCheckingMode = 'standard'
useLibraryCodeForTypes = true
disableOrganizeImports = true

[tool.ruff]
target-version='py314'

[tool.ruff.format]
quote-style = 'preserve'
indent-style = 'space'

[tool.ruff.lint]
select = ["ALL"]
preview = true
allowed-confusables = [ "–" ]
ignore = [
  "ANN003",  # missing-type-kwargs
  "COM812",  # missing-trailing-comma
  "CPY001",  # Missing copyright notice at top of file
  "D100",    # Missing docstring in public module
  "D104",    # Missing docstring in public package
  "D203",    # blank line required before class docstring
  "D205",    # blank line required between summary line and description
  "D211",    # no-blank-line-before-class
  "D212",    # summary on first line of docstring
  "D213",    # multi-line-summary-second-line
  "D405",    # capitalization in docstring section
  "D403",    # capitalization in docstring start
  "D415",    # punctuation in docstring
  "E501",    # line too long
  "EM101",   # raw-string-in-exception
  "FIX002",  # line-contains-todo
  "ISC001",  # Conflicts with formatter ruff
  "T201",    # Print found
  "TD002",   # Missing author in TODO `# TODO(<author_name>): ...`
  "TD003",   # missing-todo-link
  "TRY003",  # raise-vanilla-args
  "Q000",    # prefer double quotes
]
fixable = ["ALL"]
unfixable = [
  # "D", # dont fix docstyle from others
]
# allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"test_*.py" = [
  "S101",    # asserts allowed in tests...
  "ARG",     # Unused function args -> fixtures nevertheless are functionally relevant...
  "FBT",     # don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
  "PLR2004", # Magic value used in comparison, ...
  "S311",    # Standard pseudo-random generators are not suitable for cryptographic purposes
]
"xx_*.py" = [
  "ERA001",  # commented out code ok in scratch files
]

[tool.uv.sources]
az-pmp-utils = { git = "https://github.com/jbgreenh/az-pmp-utils" }
